src/profile_management.py
```python
import tinder_api

class ProfileManagement:
    def __init__(self, user_data_schema):
        self.user_data_schema = user_data_schema

    def manage_profile(self, user_id, bio, photos, preferences):
        """
        Update user's Tinder profile including bio, photos, and preferences.
        """
        user_data = self.user_data_schema[user_id]
        user_data['bio'] = bio
        user_data['photos'] = photos
        user_data['preferences'] = preferences
        tinder_api.update_profile(user_id, user_data)

    def optimize_profile(self, user_id):
        """
        Provide insights and recommendations for optimizing the profile to increase match potential.
        """
        user_data = self.user_data_schema[user_id]
        # Analyze user data and provide recommendations
        # This is a placeholder for the actual implementation
        recommendations = "Based on your data, we recommend you to..."
        return recommendations
```