tests/test_profile_management.py
```python
import unittest
from src import profile_management

class TestProfileManagement(unittest.TestCase):

    def setUp(self):
        self.profile_manager = profile_management.manage_profile()
        self.profile_optimizer = profile_management.optimize_profile()

    def test_manage_profile(self):
        # Assuming the manage_profile function returns a dictionary with user profile details
        result = self.profile_manager()
        self.assertIsInstance(result, dict)
        self.assertIn('bio', result)
        self.assertIn('photos', result)
        self.assertIn('preferences', result)

    def test_optimize_profile(self):
        # Assuming the optimize_profile function returns a dictionary with optimization recommendations
        result = self.profile_optimizer()
        self.assertIsInstance(result, dict)
        self.assertIn('bio_recommendations', result)
        self.assertIn('photo_recommendations', result)
        self.assertIn('preference_recommendations', result)

if __name__ == '__main__':
    unittest.main()
```